DDF-05058 Refactors navigation-right to be extensible via plugins or extension points
