+++ ui/packages/catalog-ui-search/src/main/webapp/extension-points/navigation-right/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './navigation-right'
