+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/presentation/navigation-right/index.tsx
@@ -12,4 +12,10 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
+// export {
+//   Help,
+//   Settings,
+//   Notifications,
+//   User,
+// } from './buttons'
 export { default } from './navigation-right'
